bot/requirements.txt
python-telegram-bot[job-queue]==21.4
requests==2.31.0
python-dotenv==1.0.0
httpx
aiohttp
telethon
